+++ Source/ViewControllers/Profiles/ClassicProfileViewController.swift
@@ -4,7 +4,6 @@
 //
 
 import UIKit
-import Cartography
 import Deferred
 import BSWFoundation
